refactor: move in memory user repository to source folder
