+++ src/usecases/register-user-on-mailing-list/index.ts
@@ -1,0 +1,1 @@
+export * from './register-user-on-mailing-list'
