+++ src/usecases/register-user-on-mailing-list/ports/user-repository.ts
