+++ src/usecases/register-user-on-mailing-list/register-user-on-mailing-list.ts
@@ -1,0 +1,25 @@
+import { UserRepository } from '@/usecases/register-user-on-mailing-list/ports'
+import { User, UserData } from '@/entities'
+import { Either, left, right } from '@/shared'
+import { InvalidNameError, InvalidEmailError } from '@/entities/errors'
+import { UseCase } from '@/usecases/ports'
+
+export class RegisterUserOnMailingList implements UseCase {
+  private readonly userRepo: UserRepository
+
+  constructor (userRepo: UserRepository) {
+    this.userRepo = userRepo
+  }
+
+  public async perform (request: UserData): Promise<Either<InvalidNameError | InvalidEmailError, UserData>> {
+    const userOrError = User.create(request)
+    if (userOrError.isLeft()) {
+      return left(userOrError.value)
+    }
+
+    if (!await this.userRepo.exists(request)) {
+      await this.userRepo.add(request)
+    }
+    return right(request)
+  }
+}
