+++ src/usecases/register-user-on-mailing-list/repository/index.ts
@@ -1,0 +1,1 @@
+export * from './in-memory-user-repository'
