--- src/usecases/register-user-on-maling-list/index.ts
@@ -1,1 +1,0 @@
-export * from './register-user-on-maling-list'
